reference/netology_pd_diplom/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-rest-passwordreset==1.3.1
drf-yasg==1.21.7
celery==5.3.4
redis==4.6.0
ujson==5.8.0
PyYAML==6.0.1
requests==2.31.0